Move out screenshot engine to own file
