+++ edit.js
@@ -8,13 +8,13 @@ import {TransformControls} from './TransformControls.js';
 import {XRPackage, pe, renderer, scene, camera, floorMesh, proxySession, getRealSession, loginManager} from './run.js';
 import {downloadFile, readFile, bindUploadFileButton} from 'https://static.xrpackage.org/xrpackage/util.js';
 import {wireframeMaterial, getWireframeMesh, meshIdToArray, decorateRaycastMesh, VolumeRaycaster} from './volume.js';
-import './gif.js';
 // import {makeWristMenu, makeHighlightMesh, makeRayMesh} from './vr-ui.js';
 import {makeLineMesh, makeTeleportMesh} from './teleport.js';
 
 import targetMeshGeometry from './edit/targetMeshGeometry.js';
 import attachEventListeners from './edit/classlistHandlers.js';
 import dragHandlers from './edit/dragHandlers.js';
+import screenshotEngine from './edit/screenshotEngine.js';
 import {
   worldSaveButton, worldRevertButton, micButton, dropdownButton, sandboxButton, newWorldButton,
   packagesSubpage, inventorySubpage, avatarSubpage, avatarSubpageContent,
@@ -1077,94 +1077,11 @@ for (let i = 0; i < inventorySubtabs.length; i++) {
   });
 } */
 
-async function screenshotEngine() {
-  const center = new THREE.Vector3(0, 0, 0);
-  const size = new THREE.Vector3(3, 3, 3);
-
-  const width = 512;
-  const height = width;
-  const gif = new GIF({
-    workers: 2,
-    quality: 10,
-  });
-  const gl = pe.proxyContext;
-  const framebuffer = (() => {
-    const fbo = gl.createFramebuffer();
-
-    const oldFbo = gl.getParameter(gl.FRAMEBUFFER_BINDING);
-    const oldTex = gl.getParameter(gl.TEXTURE_BINDING_2D);
-
-    gl.bindFramebuffer(gl.FRAMEBUFFER, fbo);
-
-    const colorTex = gl.createTexture();
-    gl.bindTexture(gl.TEXTURE_2D, colorTex);
-    gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, width, height, 0, gl.RGBA, gl.UNSIGNED_BYTE, null);
-    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR);
-    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR);
-    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
-    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
-    gl.framebufferTexture2D(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.TEXTURE_2D, colorTex, 0);
-
-    const depthTex = gl.createTexture();
-    gl.bindTexture(gl.TEXTURE_2D, depthTex);
-    gl.texImage2D(gl.TEXTURE_2D, 0, gl.DEPTH24_STENCIL8, width, height, 0, gl.DEPTH_STENCIL, gl.UNSIGNED_INT_24_8, null);
-    gl.framebufferTexture2D(gl.FRAMEBUFFER, gl.DEPTH_STENCIL_ATTACHMENT, gl.TEXTURE_2D, depthTex, 0);
-
-    gl.bindFramebuffer(gl.FRAMEBUFFER, oldFbo);
-    gl.bindTexture(gl.TEXTURE_2D, oldTex);
-
-    return fbo;
-  })();
-  const camera = new THREE.PerspectiveCamera(90, width / height, 0.1, 100);
-  for (let i = 0; i < Math.PI * 2; i += Math.PI * 0.05) {
-    // render
-    camera.position.copy(center)
-      .add(new THREE.Vector3(0, size.y / 2, 0))
-      .add(new THREE.Vector3(Math.cos(i + Math.PI / 2), 0, Math.sin(i + Math.PI / 2)).multiplyScalar(Math.max(size.x, size.z) * 1.2));
-    camera.lookAt(center);
-    camera.updateMatrixWorld();
-    pe.render(
-      null,
-      width,
-      height,
-      camera.matrixWorld.toArray(new Float32Array(16)),
-      camera.projectionMatrix.toArray(new Float32Array(16)),
-      framebuffer
-    );
-    // read
-    const writeCanvas = document.createElement('canvas');
-    writeCanvas.width = width;
-    writeCanvas.height = height;
-    const writeCtx = writeCanvas.getContext('2d');
-    const imageData = writeCtx.createImageData(width, height);
-    {
-      const oldFbo = gl.getParameter(gl.FRAMEBUFFER_BINDING);
-      gl.bindFramebuffer(gl.FRAMEBUFFER, framebuffer);
-      gl.readPixels(0, 0, width, height, gl.RGBA, gl.UNSIGNED_BYTE, imageData.data);
-      gl.bindFramebuffer(gl.FRAMEBUFFER, oldFbo);
-    }
-    // draw
-    writeCtx.putImageData(imageData, 0, 0);
-    // flip
-    writeCtx.globalCompositeOperation = 'copy';
-    writeCtx.scale(1, -1);
-    writeCtx.translate(0, -writeCanvas.height);
-    writeCtx.drawImage(writeCanvas, 0, 0);
-    // submit
-    gif.addFrame(writeCanvas, {delay: 50});
-  }
-  gif.render();
-
-  const blob = await new Promise((resolve, reject) => {
-    gif.on('finished', resolve);
-  });
-  return blob;
-}
 worldSaveButton.addEventListener('click', async e => {
   const {name} = pe;
   const hash = await pe.uploadScene();
 
-  const screenshotBlob = await screenshotEngine();
+  const screenshotBlob = await screenshotEngine(pe);
   const {hash: previewIconHash} = await fetch(`${apiHost}/`, {
     method: 'PUT',
     body: screenshotBlob,
